[Ingest Manager] Lift up registry/{stream,extract} functions (#83239)
## Summary

  * Move stream utility functions from `server/services/epm/registry/streams.ts` to `server/services/epm/streams.ts`
    * They're only used in registry at the moment but aren't specific to registry 
  * Move archive extraction functions from `server/services/epm/registry/extract.ts` to `server/services/epm/archive.ts`
    * The Registry isn't the only service/code which needs to extract packages. Continue consolidating archive-related code under archive vs registry
